third_party/WebKit/LayoutTests/inspector/sources/debugger/script-formatter-expected.txt

Tests the script formatting functionality.


Running: testScriptFormatterWorker

Running: testSourceMapping

Running: testFormatInlinedScripts
<html><body><script>
function f() {
}
</script><script>
function g() {
**var a;
**window.return = 10;
**if (a)
****return;
}
</script></body></html>
